latest changes to init scripts

+++ internal/CloneScriptRepo.ps1
@@ -7,8 +7,10 @@ $localRepoDir = "C:\git"
 $localScriptsRepoName = "TCCE_Scripts"
 $localScriptsRepoDir = "$localRepoDir\$localScriptsRepoName"
 
-# Remove local repo clone
-Remove-Item -Recurse -Force $localScriptsRepoDir
+# Remove local repo clone if it exists
+if (Test-Path -Path $localScriptsRepoDir) {
+    Remove-Item -Recurse -Force $localScriptsRepoDir
+}
 
 # Clone repo branch
 git clone --single-branch --branch $scriptBranch $scriptRepo $localScriptsRepoDir
